a addhousehold thunk and fetch function
