+++ api/household.ts
@@ -4,36 +4,41 @@ import {
   collection,
   deleteDoc,
   doc,
-  getDoc,
   getDocs,
   getFirestore,
   query,
+  setDoc,
   updateDoc,
   where,
 } from "firebase/firestore";
+import { generateHouseholdCode } from "../store/household/householdSlice";
 import { Household } from "../types";
 import { app } from "./config";
 
 const db = getFirestore(app);
 
-export const addHouseholdToDB = async (household: Household) => {
+export const addHouseholdToDB = async (householdName: string) => {
   const householdCollectionRef = collection(db, "households");
+  const randomCode = generateHouseholdCode();
 
   try {
-    const docRef = await addDoc(householdCollectionRef, {});
+    const docRef = await addDoc(householdCollectionRef, {
+      name: householdName,
+      code: randomCode,
+    });
 
-    household.id = docRef.id;
+    const householdId = docRef.id;
 
-    await updateDoc(docRef, household as Partial<Household>);
+    const createdHousehold = {
+      id: householdId,
+      name: householdName,
+      code: randomCode,
+    };
 
-    const householdDoc = await getDoc(docRef);
-    if (householdDoc.exists()) {
-      const householdData = householdDoc.data();
-      return householdData as Household;
-    } else {
-      console.error("Household document does not exist in the database.");
-      return null;
-    }
+    await setDoc(doc(db, "households", householdId), createdHousehold);
+
+    console.log("Household created with ID:", householdId);
+    return createdHousehold;
   } catch (error) {
     console.error("Error adding household:", error);
     return null;
@@ -41,7 +46,7 @@ export const addHouseholdToDB = async (household: Household) => {
 };
 
 export const editHouseholdToDB = async (household: Household) => {
-  console.log("hushåll som kommer in i edit: ", household)
+  console.log("hushåll som kommer in i edit: ", household);
   const householdCollectionRef = collection(db, "households");
 
   try {
@@ -138,7 +143,6 @@ export const checkHouseholdWithCode = async (joinCode: string) => {
     return null;
   }
 };
-
 
 // import { addDoc, collection, getFirestore } from "firebase/firestore";
 // import { app } from "./config";
